Move rtools.hpp to src/
