+++ src/rtools.hpp
